fix(railway): keep the country work-list out of the ais-relay watch closure
CI caught this on the merge commit: #6218 landed on main an hour after this
branch started and gave all 41 registry services an exact watch-path closure.
ais-relay now pins 49 exact patterns including scripts/_country-stock-index.mjs,
so the readFileSync this branch added to that module put
shared/openapi-filter-param-contracts.json into the always-on relay's runtime
dependency closure with no matching watch pattern. The audit was right to fail.

Widening the relay's patterns would be the wrong fix — ais-relay uses only
CHINA_COUNTRY_STOCK_INDEX_KEY and buildCountryStockIndexSnapshotFromCloses and
never touches the 45-country work-list, so watching the enum would redeploy a
long-running service every time an unrelated country entry changes, which is
exactly what #6218's exact closures exist to prevent.

loadCountryStockIndexes moves to scripts/_country-stock-index-registry.mjs,
imported only by the seeder. _country-stock-index.mjs is now free of filesystem
reads, so the relay's closure is unchanged from main.

Regression test asserts the shared module contains no readFileSync/node:fs and
that ais-relay never imports the registry module — the next person who needs a
data file in there gets a red test instead of a red deploy audit.

Verified: tests/railway-watch-path-audit.test.mjs 78 pass / 0 fail (reproduced
red before the fix, after rebasing onto main), the country-index and cache
suites 32/0, typecheck and biome clean.

Refs #6235
